+++ pyfar/level/__init__.py
@@ -4,10 +4,12 @@ from .standard_levels import (
     equivalent_continuous_level,
     exposure_level,
     sliding_equivalent_continuous_level,
+    peak_level,
 )
 
 __all__ = [
     "equivalent_continuous_level",
     "exposure_level",
     "sliding_equivalent_continuous_level",
+    "peak_level",
 ]