+++ HISTORY.rst
@@ -11,6 +11,7 @@ Added
     - `level.equivalent_continuous_level` for the Leq (PR #948)
     - `level.exposure_level` for the sound exposure level (PR #951)
     - `level.sliding_continuous_level` for the Leq with a sliding window (PR #952)
+    - `level.peak_level` for (true) peak level detection (PR #955)
 
 Fixed
 ^^^^^